🎨 Rename batch per startRumBatch for consistency
