+++ packages/rum-core/src/transport/startRumBatch.ts
